refactor: abstract pipeline ops into engine class
